% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASQ3-package.R
\docType{package}
\name{ASQ3-package}
\alias{ASQ3}
\alias{ASQ3-package}
\title{data from the ASQ3 study from CADI-UMAG}
\description{
data from the ASQ-3 study from CADI-UMAG.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nim-ach/ASQ3}
  \item \url{https://nim-ach.github.io/ASQ3/}
  \item Report bugs at \url{https://github.com/nim-ach/ASQ3/issues}
}

}
\author{
\strong{Maintainer}: Matías Castillo Aguilar \email{matcasti@umag.cl} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

Authors:
\itemize{
  \item Cristian Núñez Espinosa \email{cristian.nunez@umag.cl} (\href{https://orcid.org/0000-0002-9896-7062}{ORCID})
}

}
\keyword{internal}
