% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gam.R
\name{gam_confounders}
\alias{gam_confounders}
\title{Plot varying random effects structure GAM model}
\usage{
gam_confounders(
  data,
  var,
  seed = 1234,
  legend = TRUE,
  var_name = NULL,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{Dataset from where to extract response and independent variables.}

\item{var}{Character. Response variable to test on gam models with varying random effects structures.}

\item{seed}{Numeric. Seed for random processes within gams.}

\item{legend}{Logical. Whether to include the figure legend.}

\item{var_name}{Character. The label for the y-axis. Default is NULL, so "Predicted score" is used.}

\item{plot}{Logical. Whether to return the plot. If FALSE, then the models with the varying random effects structures are returned.}

\item{...}{Currently not used.}
}
\description{
Display graphical representation different gam models on the same
response using different random effects structure and comparing
the estimates.
}
