% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gam.R
\name{near}
\alias{near}
\title{Round to nearest multiple}
\usage{
near(i, j)
}
\arguments{
\item{i}{Numeric. Number to be rounded.}

\item{j}{Numeric. Multiple to round.}
}
\description{
Internal function used inside \code{plot_gam()} function.
It takes a \code{i} number and rounds it to the near \code{j} multiple.
}
