% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gam.R
\name{plot_gam}
\alias{plot_gam}
\title{Plot GAM model of bivariate model}
\usage{
plot_gam(data, x, y, xlab, ylab, round_to = 1, ...)
}
\arguments{
\item{data}{Dataset from where to extract x and y variables.}

\item{x}{Numeric. Variable for the x-axis.}

\item{y}{Numeric. Variable for the y-axis.}

\item{xlab}{Character. Label for x in the plot.}

\item{ylab}{Character. Label for y in the plot.}

\item{round_to}{Numeric. Intervals for mean and SE calculations.}

\item{...}{Further arguments for \code{method.args} in \code{geom_smooth}.}
}
\description{
Display graphical representation of a gam model for bivariate
relationships. Model specification can be set with the ellipsis
argument (dot-dot-dot).
}
