% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_stats.R
\name{report_chi}
\alias{report_chi}
\title{Report chi-square test}
\usage{
report_chi(chi, effsize = TRUE, ci = 0.95)
}
\arguments{
\item{chi}{A chi-square object.}

\item{effsize}{Logical. Whether to report effect sizes or not.}

\item{ci}{Numeric. Confidence interval to effect size calculation.}
}
\description{
This function lets you pass a chi-square function and report it.
}
