% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_stats.R
\name{report_kruskal}
\alias{report_kruskal}
\title{Report Kruskal-Wallis chi-squared test}
\usage{
report_kruskal(formula, data, effsize = TRUE, ci = 0.95, ...)
}
\arguments{
\item{formula}{A formula specifying the a analysis.}

\item{data}{The dataset from where to search the variables.}

\item{effsize}{Logical. Whether to report effect sizes or not.}

\item{ci}{Numeric. Confidence interval to effect size calculation.}

\item{...}{Other parameters passed to \code{kruskal.test}.}
}
\description{
This function lets you run a Kruskal-Wallis chi-squared test and report it.
}
