% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_stats.R
\name{report_overall}
\alias{report_overall}
\title{Report GAM model overall slope}
\usage{
report_overall(model, ..., ci = 0.95)
}
\arguments{
\item{model}{Object of class "gam".}

\item{...}{Further parameters to the \code{estimate_sloples()} function from \code{modelbased} package.}

\item{ci}{Numeric. Confidence interval for coefficients. Defaults to 0.95.}
}
\description{
Report the overall slopes for a GAM model fitted through the \code{mgcv} package.
}
