% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_stats.R
\name{report_pairwise_wilcox}
\alias{report_pairwise_wilcox}
\title{Report pairwise Wilcoxon rank sum test}
\usage{
report_pairwise_wilcox(formula, data, ...)
}
\arguments{
\item{formula}{A formula specifying the a analysis.}

\item{data}{The dataset from where to search the variables.}

\item{...}{Other parameters passed to \code{report_wilcox}.}
}
\description{
This function lets you run a pairwise Wilcoxon rank sum test and report it.
}
