% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_stats.R
\name{report_slopes}
\alias{report_slopes}
\title{Report linear slopes from a GAM model}
\usage{
report_slopes(model, term, ci = 0.95, length = 100, k = 1, ...)
}
\arguments{
\item{model}{Object of class "gam".}

\item{term}{Character. Term present in the model for which estimate an effect}

\item{ci}{Numeric. Confidence interval for coefficients. Defaults to 0.95.}

\item{length}{Numeric. This arguments controls the number of (equally spread) values that will be taken to represent the continuous variable.}

\item{k}{Numeric. Number of decimal places when printing numeric variables (Defaults to 1). Except for expression column.}

\item{...}{Further parameters passed to the \code{estimate_sloples()} from \code{modelbased} package.}
}
\description{
Report the linear combinations (i.e., slopes) for a GAM model fitted through effect derivatives.
}
