% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_stats.R
\name{report_smooth}
\alias{report_smooth}
\title{Report smooth terms from a GAM model}
\usage{
report_smooth(model, ...)
}
\arguments{
\item{model}{Object of class "gam".}

\item{...}{Currently ignored.}
}
\description{
Get the text format of the statistics from the smooth term
fitted from a GAM model fitted using the \code{mgcv} package.
}
